remove boom head

+++ all-silent.sh
@@ -1,4 +1,4 @@
-﻿if [[ $1 == "restart" ]]; then
+if [[ $1 == "restart" ]]; then
 	echo "restarting"
 	rm -f "$(realpath .)/buildstderrlogs.txt"
 	echo "restart done"
